move utils.io.atomic_writing to contents.fileio
